Redirect `RelayExpr` to `Expr`
As `relax.Expr` is just an alias of `tvm.relay.Expr`, but it is not
expected to be shown in the function annotation. This PR redirects
`RelayExpr` to `Expr` via customizing the hook.

Besides, `noindex` does not work for a single class, but skip the whole
module. This PR also fixes this issue.